+++ sparv/core/log_handler.py
@@ -6,6 +6,7 @@ import datetime
 import logging
 import logging.handlers
 import pickle
+import queue
 import re
 import socketserver
 import struct
@@ -305,6 +306,22 @@ class ProgressWithTable(progress.Progress):
             yield table
 
 
+class QueueHandler(logging.Handler):
+    """Custom logging handler that stores log records in a queue.
+
+    This is used when Sparv is run as a library rather than a command-line tool.
+    """
+
+    def __init__(self, log_queue: queue.Queue) -> None:
+        """Initialize handler."""
+        super().__init__()
+        self.log_queue = log_queue
+
+    def emit(self, record: logging.LogRecord) -> None:
+        """Store log record in the queue."""
+        self.log_queue.put(self.format(record))
+
+
 class LogHandler:
     """Class providing a log handler for Snakemake."""
 
@@ -320,7 +337,8 @@ class LogHandler:
         pass_through: bool = False,
         dry_run: bool = False,
         keep_going: bool = False,
-        json: bool = False
+        json: bool = False,
+        log_queue: queue.Queue | None = None,
     ) -> None:
         """Initialize log handler.
 
@@ -334,6 +352,7 @@ class LogHandler:
             dry_run: Set to True to print summary about jobs.
             keep_going: Set to True if the keepgoing flag is enabled for Snakemake.
             json: Set to True to enable JSON output.
+            log_queue: Queue to store log messages.
         """
         self.use_progressbar = progressbar and console.is_terminal
         self.simple = simple or not console.is_terminal
@@ -341,6 +360,7 @@ class LogHandler:
         self.dry_run = dry_run
         self.keep_going = keep_going
         self.json = json
+        self.log_queue = log_queue
         self.log_level = log_level
         self.log_file_level = log_file_level
         self.log_filename = None
@@ -399,8 +419,10 @@ class LogHandler:
             )
         )
 
-        # stdout logger
-        if self.json:
+        # stdout logger or log queue
+        if self.log_queue:
+            stream_handler = QueueHandler(self.log_queue)
+        elif self.json:
             stream_handler = logging.StreamHandler()
         else:
             stream_handler = ModifiedRichHandler(enable_link_path=False, rich_tracebacks=True, console=console)